Move convertPreviewToImage to Mint process
